Preceed operation response with initial put (#4)
When executing an action, the created operation could be returned with a
client response for an operation that has not been registered through
putOperation (and watchers/outstanding map). Since there must be an
entity response for watchers who request initial state, this must be
done prior to providing an endpoint to watch.