Retain Done operations for getOperation/watchers
Provide initial watcher data and a getOperation response for operations
which have completed. Expiration of this data will be left up to the
concrete instance implementation, with the deleteOperation request now
supported for completed operations. Added an operationLock abstract
method that must provide a consistent synchronization primitive for a
given operation in order to guard against race conditions when
transferring/retrieving/removing operations.