Refactor worker cache, impose size-expiration
Use a CAS-fetching file cache supplier of Paths in the worker, which
manages expiration based on a configurable maximum number of bytes. The
file cache refreshes its access time based on each action execution, and
as such is capable of reporting actual input edge traversals, as well as
retaining those file legitimately used most recently - benefit may be
negligible in this case, since inputs must be fetched even if unused,
but may aggregate over distinct actions to be effective. The file cache
will allow temporary oversubscription in this case, when all of the
inputs combined for a single action resolve to be larger than the max,
but it will prevent a single input that is larger than the max from
being inserted.