Configurable symlink of input-only directories
Provide the option of symlinking directories which are present in the
input tree and have no intersection with output file paths. This results
in substantial throughput improvement for the IF->EX transition,
allowing extremely large input trees to be coherently factored out of
the input fetch stage. Invalidation of expired content of a directory
occurs recursively in the CASFileCache.