Mitigate stalls during match stage (#127)
Matches are subject to failures on the fetch method which must not
prevent the continued operation of the stage. Guard against invalid
content in the match callback by releasing the subsequent stage, and
test for missing action blobs explicitly.
Adding tests to validate a failed fetch can continue to a second fetch
undeterred, having released the following stage that it claims again in
order to prevent overfetching.