Fix operation paging, reinforce with tests
The memory instance would not progress beyond the initial page due to an
incorrect test against the token itself rather than the deserialized
OperationIteratorToken's name, and the paged test was initialized to
true, making the paging loop break on the first entry anyway.

Added tests to guard operation list retrieval.