Inhibit empty file put into CASFileCache
Refactor and converge some of the methods surrounding CASFileCache::put
invocations and prevent empty files from being inserted and reference
counted. Empty files suffice instead of hard links to the file content.
This also prevents any attempts to fetch empty digests via the
inputStreamFactory.

Simplified public interface to 'putDirectory', and 'putDirectoryFiles'.
'put' is now an implementation detail and the containingDirectory
parameter to it is defaulted.