Remove the worker container image.
The container image for the worker process that we currently ship is
effectively useless. It's based on a distroless container image, meaning
it doesn't even include a shell interpreter. Because of this, it is not
able to spawn any subprocesses.

People who want to run workers through Docker should create their own
container images with the development tools they need.