Provide ExecutionPolicy specification on workers
Workers may declare a series of execution policies that can be referred
to by platform properties of name `execution-policy`. The available
policy is an ExecutionWrapper, where a path string will be concatenated
with the current command arguments to be executed. These policies can be
specified multiple times in the platform properties, where later
wrappers will wrap earlier ones.