Provide ExecutionPolicy specification on workers (#188)
Workers may declare a series of execution policies that can be referred
to by platform properties of name `execution-policy`. The available
policy is an ExecutionWrapper, where a path string will be concatenated
with the current command arguments to be executed. These policies can be
specified multiple times in the platform properties, where later
wrappers will wrap earlier ones.

A default execution policy in the worker config with an empty name will apply
universally to all executions if specified, before any other execution policy is
applied.