Remove findMissingBlobs from getActionResult
All clients support treatment of missing results in the CAS with a
remote execution. This will drastically reduce the cost of delivering
cached results.

Force duplicate executes by metadata to skip cache

Prevent a sequence of re-executions of the same action, determined by
RequestMetadata, from being served from the cache by forcing them to
skip the lookup after they have been served from it. This compensates
for a behavior by bazel which does not modify this flag after it has
failed to download a blob in the results when it was missing.