Provide server-side keepalive for operations
Issue a keepalive operation on a configurable period indicating that the
work is still outstanding and to mitigate any network idleness issues
along the client route. If left unconfigured, or <= 0, no keepalives
will occur.

Keepalives will be cancelled with the execution context, and will not be
delivered or rescheduled if the context is cancelled.