Sequence and accommodate for async upload complete
Ensure that responses to Write are sequenced through synchronization
with the WriteStreamObserver.

WriteRequests that are received after a commit through asynchronous
completion are ignored.

Stub Writes that have been completed ignore resets and are immediately
completed with the final committed size. OutputStream operations after a
completion can throw a received error from the stub.