Rework correctMissingBlob for completions size
Specify a completion count for an aggregate callback, since the number
of completions is a fixed size. The prior implementation was observed to
complete after a single response was received, confirmed through
logging. This has exhibited no such behavior and has corrected for a
series of blobs during downgrades and through multiple builds.