Institute a configurable max blob size for writes
Provide a configurable value for the maximum size of an attempted blob
upload. Specifying <= 0 for this value makes blob size unlimited. This
limit operates only at the ShardInstance, and does not affect output
uploads, which effectively limits large outputs to the executing
workers. RequestMetadata is propagated to the blob write creation,
enabling attribution when it occurs and is logged. Clients will observe
this limit being exceeded as an UNAVAILABLE grpc status in order to
facilitate client fallback.