Reverse order of shutdown sequence in OQ worker
The Superscalar stage will wait indefinitely while claimed for a
release, assuming that it is held by a running worker. Since the
MatchStage holds a claim while waiting for a match to prevent
oversubscription, the InputFetchStage will not shut down while a match
is waiting. Interrupting the match stage at highest priority, and
downstream stages in descending order of the pipeline allows a full
flush in ideal circumstances.