Provide stderr text when process start fails
If a ProcessBuilder#start fails, present a reasonable message to
indicate via the operation what occurred. Since openjdk's default
start IOException wrapper includes a directory that will be meaningless
on the client, attempt to intercept it by expecting a cause, which
should only be the literal failure from exec, and constructing a simple
decoration of the command.