Fix external import and instructions
Label references to resources within the external repo must be fully
qualified with the repository name. If someone really wants to change
the name, they can do so by passing `repository_name` to our deps
method. Moved the googleapis and remote_apis build definitions to
third_party, with the hope that we can remove these entirely.