Reflect empty executable bit, directory compute
The empty executable bit must be reflected from a FileNode which
requests it, because when the CASFileCache computes its directory
entries, it must see that bit in order to successfully rediscover the
directory hash for validation.

This also highlighted a bug in computeDirectory, where no subdirectories
could ever be considered for recursion, due to a mismatched path check.