Put maxTimeout validation in impl for shard
The maximum timeout validation override must match the implementation
used for actions. This properly enables scheduler rejection, including
requeue.

Note that a mismatched timeout on workers will result in operation
looping due to non-terminal operation status - a worker does not know
that an action is definitively invalid for all cases.