Properly handle wrapper argument substitutions.
When encountering a wrapper argument of the form <propertyName>, the
Executor will currently add both the value of the property named
propertyName as well as the original string '<propertyName>'. This seems
incorrect.

This change makes it so that the executor will only add the property
value corresponding to propertyName and not the original argument string
'<propertyName>'