Wait/interpret NOT_FOUND in ByteStreamHelper (#431)
In order to properly failover from the local shard CAS to remote, the
FailoverInputStreamFactory must observe NoSuchFileException on
InputStream creation for the primary. This comes into play when a
CASFileCache is delegating to a grpc CAS as the primary.

An eventual target here may be the return of a modified InputStream that
can failover at any point in its lifecycle, so that blob disappearances
can be compensated for over multple requests, due to deadline
expiration or transiency.