Retain a single FileStore for a cfc root
Under the assumption that a cfc root does not cross filesystems, use a
single reference to the fileStore to make util decisions (this could
also be done by providing a specific IOUtils instance per-fs-type).

The FileStore's computation was observably painful and appears to be
parsing a mountset on each invocation of Files.getFileStore.