Serialize bytestream proxy callbacks to a single thread in StubInstance.
This commit moves all interaction with the server response stream and
client stub onto a single thread. This allows us to have guarantees
that onNext and onReady will not execute in an interleaved fashion.
Critical sections could have been added, but blocking either the client
stub to the shard worker or server stream could have impacts on
performance of other streams being handled on those transports.