Restore stub read impl, refactor, test
The recent modifications to guard against a race between isReady and
wasReady for stub getBlobs left the possibility of an onReady call from
a blobObserver before the requestStream was started. Calling
`request(1)` on this stream would result in an invalid state and ensuing
exception.

This change moves the class declaration to a static member scope of
StubInstance, removes the required ExecutorService member, simplifies
the state tracking to remove atomics, and clarifies guards for accessing
the state indicators. wasCompleted is left as it represents a unique
volatility, but might not be needed as a wrapper.

The constraints of the original issue #440 are satisfied here, as well
as some additional testing for this scenario, where a request(1) does
not occur without an onNext (the only guarantee that a client call has
started) call.