Prevent error flood on removed entry during expire
Entries appear to be capable of being removed from storage, but not
unlinked, when they reach the head of the LRU for expiration. Guard
against this loop by unlinking an entry if it expresses linkage.

This is not a correction for the inconsistency, which appears to be
guarded correctly, but for avoiding the unresolvable state of the same
entry being removed from the LRU list.