Restore aggregate worker blob registrations
The aggregated CFCExecFileSystem start of CFC was intended to provide an
aggregate mechanism for advertising blobs. This was subverted when a
refactor happened for CFC startup, and never corrected. The onPut
mechanism is pretty awkward to begin with, and should be a general
purpose notification system. Refactor in CFC is becoming a necessity.

Changed the variable name for the onStartPut callback, but we should
really prevent the start routines from seeing the onPut at all through
more reliable methods.