Document Redis timeout and maxAttempts settings
This is related to https://github.com/bazelbuild/bazel-buildfarm/pull/583

In case a Redis node is busy, the short timeout settings (2 seconds)
can make the cluster member resolution routine fail, so the buildfarm
process can end prematurely. If the buildfam process is immediately
restarted, it can overload the same node again, creating a start loop.

While we worked around the problem by modifying the source code,
the configuration settings are already present and will be
made more visible with this pull request.