Limiting to a single active CFC output stream (#662)
Races with concurrent writes due to buffering of WriteRequests must not
be allowed to introduce corrupted stream data. Prevent any more than a
single outstanding output stream before close in order to ensure
serialized writes and a committedSize check. This is implemented as
blocking for now in the WriteStreamObserver, but a future interface is
also presented and tested. WriteStreamObserver now also is willing to
skip ahead in WriteRequests until a committedSize is reached by offset,
to help mitigate the cost of buffered inconsistencies (the stream makes
it further since the interleaved QueryWriteStatus due to outstanding
operations).

This exposed several problems with size accounting in writes, including
a charge for each attempted write that received data for the full
amount, as well as successive filesystem queries and isReset logic
failures for request recycling and resumption.

The cost to recalculate the stream digest in order to resume becomes
dominant at short deadlines. A further improvement would be to take
steps to see that the hash computation is recorded with the write while
it remains in the live write list.