Propagate IOException without delegate on newInput
The failover (local->remote) input stream factory relies on the specific
NoSuchFileException presentation by the CAS in order to trigger a remote
(other shard member) download for the worker context getBlob.
newInputFallback must not be called when delegate is null, or it will
incur an NPE from the precondition validation.