Consistently apply empty blob directory policy (#1106)
A pass through a directory hierarchy may encounter multiple empty blob
directories under the root for a consistent hash which does not appear
in the directories index - it may be empty or any other value upon
computation. This change ensures the secondary pass after visited
properly handles the empty digest consistent with the first behavior
which uses a default instance. This is a likely cause of NPEs in the
secondary pass, not guaranteed to be empty directories, but observably
caused by them in the unfixed test case.

Fixes #1101