Apple based execution wrapper
Implement an execution wrapper for macOS to set `SDKROOT` and `DEVELOPER_DIR`

In Apple based builds - this is required by clang, llvm, and related
native compilers and linkers.

Because there is no way to govern where Xcode.app is installed locally
or remotly it must be detected by Bazel. Because isn't possible to
assume the remote path is identical to the local path a buildfarm worker
must solve for this. We either query launchd where it was installed (
using Bazel's `xcode-locator` on darwin) _or_ map to the directory based
on convention. I added documentation about this problem and how it's
solved in the wrapper incase folks are new to this.

At the time of writing there is no notion of a macOS worker in this repo
- so I didn't add build rules for this yet. We don't have docker on
macOS but, I'd propose making a tarball rule longer term if people are
on-board with this idea. We can work towards a reference and canonical
example to build and release a worker end to end including, config, and
wrapper, and xcode-locator.