[breaking config] Convert all proto configs to yml
BREAKING CONFIG CHANGE. BREAKS COMPATIBILITY WITH PREVIOUS 1.x VERSIONS!!!

Convert protobuf configs to yaml. This PR will become a start of v2.0 and will break compatibility with v1.x.
The benefits of moving to yaml configs is to simplify configuration, remove dependency on classes on each other and open up the possibility for future refactoring and simplifying of the codebase.