Restore containsBlob functionality for shard
Needed for FetchService
FindMissingBlobs requests may contain any number of -1 sized digests
The response for -1 will be the original -1 sized digest if missing (unchanged from >0), and the digest populated with size if present. This differs from >0 behavior by returning a value for entries which exist, but functionally returns an unrequested digest in the response, differing by size.