[Worker][Resilience] Graceful term on PipelineStage failure
Today a pipeline may get into a undefined state but leave a worker in a
hung but running state. This PR gracefully terms the worker when
catching exceptions in a `PipelineStage`

Do this by means of a `ThreadGroup` which has an uncaught exception
handler. This design affords catching exceptions from all pipelines and
delegating actions to the worker, which makes an attempt to interrupt
stage threads to shut down. It necessitated removing the catch all
exception swallowing feature of the `PipelineStage`.

Currently a worker will `exitPostPipelineFailure` but I imagine folks
can extend this and provide parameter to take other actions.