Giving names to worker threads
Use a schema where the thread is named by the class that owns it and a
brief description about what it does, based on what people called the
variables.

The context is I've been trailing a bug from time to time and this makes
it easier to look at dumps. If we like the format I'd be happy to go
through and name other threads in the repo using this convention