Restore DequeueMatchSettings Platform
The DequeueMatchSettings Platform before 0a6802e82f provided for
selection or rejection of queue entries based on locally selectable
platform properties that it provides, per worker. The config transition
eliminated this capability, and left the workers only with the
properties selected from the queue definition restrictions. Highlighting
this was the elimination of use of the platform during
ShardWorkerContext test creation, leaving a platform parameter orphaned.

Use the DMS Platform both in eligibility filtering and selection of the
eligible queues to permit mildly heterogenous worker queue processing.

Fixes #1350