[Draft] Atomic locking of `createExecDir`
Summary: today we reference an `Entry` after it has been "downloaded".
Today we expire _during_ execution and `InputFetch` stage when `Entries`
for keys either existed on the file system or were loaded from the
distributed store in the `InputFetch` operation. Will also mutate
existing entries due to `deleting` the reference to the last inode of a
file. The latter manfiests in several ways e.g. like failing the report
stage of a concurrent action.

Prior to `CASFileCache.putDirectory` we increment `keyReferences` to
prevent eviction during execution. This provides atomicity gaurentees
around the _keys_ within  directories while minimizing mutexes.

Fixes my contention with `createExecDir` / and my write-to-CAS failures
which manifested as a result.

TLDR:

Logical and atomic locking of an `execDir` by gaining references to keys
via the new method `lockDirectoryKeys`. Because there are many
references to a single `Entry` this needs to be refernce counted for
each concurrent `execDir` and must upheld through the entire lifetime of
the directory.

Testing:

I have a number of manual integration tests that need reworking so I can
add them here and will add these as part of the PR to get it out of the
`Draft` state