Interpret ExecDirException to Status (#1391)
Handle ExecDirExceptions in InputFetcher such that the client can
observe a FAILED_PRECONDITION with PreconditionFailures that include
Violations with VIOLATION_TYPE_MISSING to inspire virtuous loop
reestablishment. A ViolationException acts as a container for this, and
can be extended to include the components of a putDirectory. Interpret
PutDirectoryExceptions with their imposed violations. Missing inputs at
the time of execute, whether linked as immediate files or through the
directory cache, will be interpreted as VIOLATION_TYPE_MISSING.