[Server][RFC][WIP] Persist write instances in redis
We see some flake in the system which ends up resulting in a number of
byte stream requests routing to the wrong CAS worker. These problems are
rooted in both load balancing assumptions and scheduler failures. One
problem with trying to resume writes after failures is being able to
find the instance it was writing to: we see `write` calls failing across
all cas workers.

Testing it out for the problem cases we see. Error conditions: if the
write instance goes out it is still persisted: now both redis and in the
`LoadingCache`.

This has a lower timeout than the loading cache - to prevent a large
expansion of redis usage. A better alternative to hard coding these would
configure the write timeout in the config file, but skipping for now.