Write onNext requires onCompleted to follow (#1565)
For reasons passing understanding, a WriteResponse is not delivered to a
client solely from onNext to be received by a client called via bsStub.
onCompleted must follow the onNext for a timely delivery of the
response, particularly when it occurs before an onCompleted from the
client. This does not seem to affect the bazel client, which could
indicate a difference in behavior between grpc versions, or simply that
it manages to complete its upload completely ignorant of the early
return status. More investigation is required.

This will fix hanging builds which occur with exec-only workers failing
to complete uploads to shard peers.