Dequeue flags expire after timeout effect (#1642)
The flag key which contains a timeout for an operation that has been
observed in a Queue's Dequeue (atomic safeguard list, either queue or
dequeue) must not expire before the time it will be expired, and
reasonably not before the time it takes a failsafe period to identify
the timeout has elapsed. This was broken in #713, where the original
expiry was 2x the configured timeout, and reduced to just the timeout in
seconds. We now set the expiry to 1 hour or 10x the timeout, to make
unlikely an expiry clearing the flag in a working system. Commented
loudly to avoid removing the same padding again.