publish Helm chart as OCI artifact to ghcr.io (#1683)
in lieu of publishing a tgz as a GitHub release artifact, this change publishes (on tag helm/${semver}) the Helm chart bundled as an OCI artifact to the GitHub container registry, i.e. ghcr.io. I also updated the installation instructions which are now less gory. This change requires granting public read access to the corresponding buildfarm "package" within the bazel-buildfarm project. It defaults to private.