Limit Zstd ByteBuffer allocations with BufferPool
Implement a limiting BufferPool for Zstd's decompression sources.
This may mitigate OOMs which occur as a result of many concurrent
writes, but does nothing to capitalize on commonly short compressed
blobs.
Since there is no control over the buffered read size at the zstd-jni
level, lower-level wrappers must be written to take advantage of this,
and short-blob short-circuiting in CASFileCache may be necessary as
well.