Extend NSFE removal to all IOException (#1773)
Any IOException that results in a file being inaccessible for read now
triggers a storage removal.

There is the off chance that there are exception conditions that are
process resident (too many files open) that are transient where this may
go too far. Unfortunately there is no great distinction between these
exceptions and those that are unrecoverable based on current state. To
be safe (because safety is convergence) we will attempt a full removal
of the afflicted file, and survive any remaining IOExceptions with
suppression under the primary one.

Getting the filesystem to kick out the observed FileSystemException that
occurs when, say, a kernel blk device report has occurred and a mounted
xfs has denied access to the filesystem is problematic. The existing
missing result is verified through test `readRemovesNonexistentEntry`,
albeit at a low level - this will at least begin to hit on
findMissingBlobs and prevent this worker from crippling a cluster due to
a failure to converge.

We add a read_io_errors counter and log the unexpected error in all but
the NSFE case, which is a functionally acceptable state to be in.