allowedUnmatched is true sooner
We have an unusual bug in our buildfarm deploy. Some product-teams send us "OSFamily" and "Linux" and this does NOT match "OSFamily" and "Linux" in the configuration used by servers or workers. 

Changing to use "*" as the OSFamily doesn't work either. We change the logic to avoid checking the properties and the situation resolves.

Is this as intended? Should `allowUnmatched: true` refuse to match when the properties don't match?