POOL Resource type
Segment resource types into semaphores and pools
Pools create a queue of ids in the range [0,amount)
These ids are depleted and replenished via resourced executions
ExecutionPolicies with pool-<resourceName> are automatically effective
when a resource is requested.
ExecutionWrappers should interpolate the resource name in a parameter to
adapt the allocated ids to more specific resource isolation mechanisms.

The MatchListener now delivers a Claim, which must be released by
consumers to replenish used pool resources.
The error pipeline path for Match and InputFetch stages now release
claims in the ExecutionContext and requeue an operation.

Minor items for concern:
ExecutionContext was not previously available to the worker.shard
namespace, and providing a safer error path that includes a release and
requeue hs required this. Further refactoring/context expansion may be
necessary.
The Executor interprets LocalResourceSetClaim as the only possible
presentation of ExecutionContext.Claim.

Additionally, resources now extend the overall satisfaction of execution
properties automatically (with any value, subject to claim satisfaction
with depletion), unless the dequeueMatchSettings properties specify
resource:<name> with an explicit value, where this must be matched
exactly, as before.