Only stop WorkerStub instances on idle
Monitor the ConnectivityState for WorkerStubs' channels.
When all channels go idle, the instance will be stopped.
This will prevent rejected executions when a retry executor has been
stopped, while still bounding the instance lifetime, and permitting
expiration in the worker stub cache to ensure limited channel use.

All callers are required to register onStopped callbacks which
invalidate against the Cache - a clear solution for a value-state-change-
initiated action against the LoadingCache object is not obvious without
hooking the response. Could be reorganized to another static method in
WorkerStubs, or turning WorkerStubs instance into a subclass of cache.