Simplify and configurize windows-default command handling
Windows requiring an absolutized path does not apply correctly to things
like genrule, which will invoke a command with `cmd.exe ... <program>`.
Absolutization of the command is a transform that now checks to see
whether the path exists relative to the working directory, permitting
both in-tree and system command invocation.
This also cleans up the iteration and pushes the os.name invariant
interpretation to configs, where it can be overridden if desired.