feat: support cgroups v2 and v1 (#1884)
* feat: support cgroups v2 and v1

This adds support for CGroups v2 and maintains support for CGroups v1
for one more release.

Efforts made to keep v1 and v2 separate so that v1 support can be
removed without a ton more refactoring.

* tests(cgroups): fix tests for Windows

* feat: evacuation

This sets up two cgroups under the group where we were launched:
- one called "evacuation" where all current PIDs go.
- it makes room for the "executions" CGroup.

- "evacuation" and "executions" are now siblings in the cgroup
hierarchy, and no processes remain in the cgroup where we were launched.

* feat: cgexec replacement

* chore: default cgexec-wrapper

* feat: better v1 detection support

* feat: support two execution wrappers

* debug: Warning if there's NO cgroups at all

* fix(cgroups): implement Cpu.setShares(int)

* feat: cgroups v2 shutdown hook

* fix: catch more exception classes

* fix: support start-up for v1 and v2 better

* fix: don't set "tasks" owner for cgroups v2

* doc notes

* tests(cgroups): Move logic into CGroupVersionProvider and test it

Adding tests

* tests(cgroups): tests for getSelfCgroup

* chore(license): adding license headers

* refactor: use a constant for the evacuation group name

* fix(cgroups): avoid nesting dolls of evacuation cgroup hierarchy

If we're starting up in a cgroup ending in "evacuation", pick the parent
path (and log it).

* doc(cgroups): add javadoc

* fix: cleaner startup when booting in evacuated group

* fixup! fix: cleaner startup when booting in evacuated group