Zstd committed size (#2251)
* Present uncompressed committedSize for queries

Ensure that committedSize is presented as the uncompressed size for
stream restarts. Only check the committedSize for successive requests
against the originally requested offset + the compressed bytes written
in a stream.

* writeFuture.isDone() is definitively 'completed'

Attempts to decide between the 'correct' size response for a write are
shortsighted. For combined execution/output stream types we have:
Long.MAX_VALUE for 'unlimited' streams, which defer to future
-1 for compressed streams, which have undefined requirements for QWS
The expected blob size, which doesn't strictly mean complete.

If a QWS returns 'completed', the write must always present as
isComplete. All other writes are incomplete.
This also means that a blob write can remain at the given offset and not
be complete. Whether this should return -1 or the full size (or neither)
is undefined by API documentation.