Include the exec owner in the work root hash
Because the exec owner is now part of the worker key, it's possible to
create workers that are identical but whose worker keys are otherwise
identical. This is a problem because if one worker is torn down before
another, the other will have its exec root swept from under its feet,
causing weird errors about files failing to copy, missing files, etc.

To resolve this, we need to include the two fields we previously added
to `WorkerKey` (`owner` and `wrapperArguments`) when generating the
hash.