Handle CancellationException from input fetch
A CancellationException thrown during the fetch will result in a failure
to wait for all subsequent fetch completions. Enumerate this exception
in the list and continue to ensure that the finally recovery which
decrements references in the CAS runs without possibility of concurrent
modification, detected or oblivious, crashing the recovery or leaking
references, respectively.

Fixes #2411